+++ hoomd/hpmc/pair/zetterling.py
@@ -28,18 +28,14 @@ class Zetterling(Pair):
         mode (str): Energy shifting/smoothing mode.
 
     `Zetterling` computes the oscillating pair potential between every pair of
-    particles in the simulation state. The functional behavior of the potential
-    under the various shifting modes is the same as in `hoomd.md.pair`.
+    particles in the simulation state. The functional form of the potential,
+    including its behavior under shifting modes, is identical to that in
+    the MD pair potential `hoomd.md.pair.Zetterling`.
 
-    .. math::
-        U(r) = A \\frac{\\exp{(\\alpha r)\\cos{(2 k_F r)}}}{r^3}
-              + B \\left( \\frac{\\sigma}{r} \\right)^n
+    See Also:
+    `hoomd.md.pair.Zetterling`
 
-    The potential was introduced in `F. H. M. Zetterling, M. Dzugutov, and S. Lidin
-    2001`_.
-
-    .. _F. H. M. Zetterling, M. Dzugutov, and S. Lidin 2001:
-       https://doi.org/10.1557/PROC-643-K9.5
+    `hoomd.md.pair`
 
     .. rubric:: Example
 
@@ -55,6 +51,7 @@ class Zetterling(Pair):
             "n": 18.0,
             "r_cut": 2.649,
         }
+        simulation.operations.integrator.pair_potentials = [zetterling]
 
     {inherited}