+++ hoomd/md/module-md.cc
@@ -67,10 +67,12 @@ void export_PotentialPairLJGauss(pybind11::module& m);
 void export_PotentialPairForceShiftedLJ(pybind11::module& m);
 void export_PotentialPairTable(pybind11::module& m);
 void export_PotentialPairWangFrenkel(pybind11::module& m);
+void export_PotentialPairZetterling(pybind11::module& m);
 
 void export_AnisoPotentialPairALJ2D(pybind11::module& m);
 void export_AnisoPotentialPairALJ3D(pybind11::module& m);
 void export_AnisoPotentialPairDipole(pybind11::module& m);
+void export_AnisoPotentialPairYLZ(pybind11::module& m);
 void export_AnisoPotentialPairGB(pybind11::module& m);
 void export_AnisoPotentialPairPatchyExpandedGaussian(pybind11::module& m);
 ;
@@ -81,6 +83,8 @@ void export_AnisoPotentialPairPatchyLJ(pybind11::module& m);
 void export_AnisoPotentialPairPatchyMie(pybind11::module& m);
 void export_AnisoPotentialPairPatchyYukawa(pybind11::module& m);
 void export_AnisoPotentialPairPatchyTable(pybind11::module& m);
+
+void export_MeshForceCompute(pybind11::module& m);
 
 void export_PotentialBondHarmonic(pybind11::module& m);
 void export_PotentialBondFENE(pybind11::module& m);
@@ -116,6 +120,8 @@ void export_PotentialExternalWallMorse(pybind11::module& m);
 
 void export_PotentialPairDPDThermoDPD(pybind11::module& m);
 void export_PotentialPairDPDThermoLJ(pybind11::module& m);
+
+void export_MeshDynamicBondUpdater(pybind11::module& m);
 
 void export_IntegratorTwoStep(pybind11::module& m);
 void export_IntegrationMethodTwoStep(pybind11::module& m);
@@ -227,10 +233,12 @@ void export_PotentialPairLJGaussGPU(pybind11::module& m);
 void export_PotentialPairForceShiftedLJGPU(pybind11::module& m);
 void export_PotentialPairTableGPU(pybind11::module& m);
 void export_PotentialPairConservativeDPDGPU(pybind11::module& m);
+void export_PotentialPairZetterlingGPU(pybind11::module& m);
 
 void export_AnisoPotentialPairALJ2DGPU(pybind11::module& m);
 void export_AnisoPotentialPairALJ3DGPU(pybind11::module& m);
 void export_AnisoPotentialPairDipoleGPU(pybind11::module& m);
+void export_AnisoPotentialPairYLZGPU(pybind11::module& m);
 void export_AnisoPotentialPairGBGPU(pybind11::module& m);
 
 void export_AnisoPotentialPairPatchyExpandedGaussianGPU(pybind11::module& m);
@@ -367,6 +375,7 @@ PYBIND11_MODULE(_md, m)
     export_PotentialPairForceShiftedLJ(m);
     export_PotentialPairTable(m);
     export_PotentialPairWangFrenkel(m);
+    export_PotentialPairZetterling(m);
 
     export_AlchemicalMDParticles(m);
 
@@ -379,6 +388,7 @@ PYBIND11_MODULE(_md, m)
     export_AnisoPotentialPairALJ2D(m);
     export_AnisoPotentialPairALJ3D(m);
     export_AnisoPotentialPairDipole(m);
+    export_AnisoPotentialPairYLZ(m);
     export_AnisoPotentialPairGB(m);
     export_AnisoPotentialPairPatchyExpandedGaussian(m);
     export_AnisoPotentialPairPatchyExpandedLJ(m);
@@ -391,6 +401,8 @@ PYBIND11_MODULE(_md, m)
 
     export_PotentialPairDPDThermoDPD(m);
     export_PotentialPairDPDThermoLJ(m);
+
+    export_MeshForceCompute(m);
 
     export_PotentialBondHarmonic(m);
     export_PotentialBondFENE(m);
@@ -466,6 +478,7 @@ PYBIND11_MODULE(_md, m)
     export_PotentialPairTableGPU(m);
     export_PotentialPairWangFrenkelGPU(m);
     export_PotentialPairConservativeDPDGPU(m);
+    export_PotentialPairZetterlingGPU(m);
 
     export_PotentialTersoffGPU(m);
     export_PotentialSquareDensityGPU(m);
@@ -477,6 +490,7 @@ PYBIND11_MODULE(_md, m)
     export_AnisoPotentialPairALJ2DGPU(m);
     export_AnisoPotentialPairALJ3DGPU(m);
     export_AnisoPotentialPairDipoleGPU(m);
+    export_AnisoPotentialPairYLZGPU(m);
     export_AnisoPotentialPairGBGPU(m);
 
     export_AnisoPotentialPairPatchyExpandedGaussianGPU(m);
@@ -544,6 +558,7 @@ PYBIND11_MODULE(_md, m)
     export_MTTKThermostat(m);
     export_BussiThermostat(m);
     export_BerendsenThermostat(m);
+    export_MeshDynamicBondUpdater(m);
 
     export_IntegratorTwoStep(m);
     export_IntegrationMethodTwoStep(m);