+++ hoomd/md/pair/__init__.py
@@ -146,6 +146,7 @@ from .pair import (
     ForceShiftedLJ,
     Moliere,
     ZBL,
+    Zetterling,
     Mie,
     ExpandedMie,
     ReactionField,
@@ -189,5 +190,6 @@ __all__ = [
     "Table",
     "WangFrenkel",
     "Yukawa",
+    "Zetterling",
     "aniso",
 ]