+++ CHANGELOG.rst
@@ -4,6 +4,25 @@ Change Log
 5.x
 ---
 
+5.4.0 (2025-09-26)
+^^^^^^^^^^^^^^^^^^^^
+
+*Added*
+
+* The Zetterling MD pair potential: ``hoomd.md.pair.Zetterling``
+  (`#2086 <https://github.com/glotzerlab/hoomd-blue/pull/2086>`__).
+* The YLZ anisotropic MD pair potential: ``hoomd.md.pair.aniso.YLZ``
+  (`#2110 <https://github.com/glotzerlab/hoomd-blue/pull/2110>`__).
+* Dynamic bond topologies with mesh potentials: ``hoomd.md.update.MeshDynamicalBonding``
+  (`#2093 <https://github.com/glotzerlab/hoomd-blue/pull/2093>`__).
+
+*Fixed*
+
+* Display vector math symbol correctly in the documentation
+  (`#2109 <https://github.com/glotzerlab/hoomd-blue/pull/2109>`__).
+* Display equations correctly in the documentation
+  (`#2118 <https://github.com/glotzerlab/hoomd-blue/pull/2118>`__).
+
 5.3.1 (2025-07-18)
 ^^^^^^^^^^^^^^^^^^